web-app/requirements.txt
Django>=3.0.7
mysqlclient>=1.4.6
django-mysql>=3.6.0
djangorestframework>=3.11.0
fcm-django>=0.3.4